# openapi.yaml
openapi: 3.1.0
info:
  title: Contract analyzer API
  version: 1.0.0
  description: |
    Wire contract for the three-arm Polish contract analysis artefact.

    OWNERSHIP: this document owns the wire contract.
    tests/backend/test_api.py::test_openapi_conformance asserts that the FastAPI-generated
    schema conforms to this file; drift is a test failure, not a runtime surprise.

    DERIVED FROM: src/contract_analyzer/domain.py (frozen records and enums). Response
    schemas mirror the domain records; they are not invented here.

    THE THREE ARMS differ only in the call-unit constructor. Model, prompts, corpus,
    tools, retry policy, concurrency and budgets are identical across arms. A client must
    not present the arms as ordinal or as quality levels.

    PRIVACY BOUNDARY: document text lives only in process memory under an opaque document
    session and is purged on delete, session expiry, or process exit. After purge, metadata
    and metrics remain readable while content and chat return 410. SSE events never carry
    document or corpus text.

    WHERE THE ARCHITECTURE IS SILENT this document decides:
      * limit_breach maps to 413, not 422: 422 is for "malformed or empty" content,
        and a breach is neither.
      * missing_tool maps to 503: a missing converter or OCR data is a dependency failure
        that readiness already gates, not a defect in the submitted file.
...
servers:
  - url: /api/v1
...
paths:
  /documents:
    post:
      tags: [documents]
      summary: Upload and ingest one document
      description: |
        Ingests TXT, DOCX, DOC or PDF once into canonical text, anchors and reference
        records. All arms reuse these exact objects. PDF read mode is a whole-document
        decision; native and OCR pages are never mixed.

        Format is chosen by the filename extension of the multipart part (.txt, .docx,
        .doc, .pdf), not by media type. The filename is significant. One exception: a
        .doc part whose first bytes are a ZIP local file header is read as .docx. No
        other extension is read by content.

        The client must state, before the first run, that OCR runs in-process, that
        retrieval sends model-chosen search phrases, which may contain document
        excerpts, to an external embedding provider, and that the selected document
        text is sent to the model provider for analysis.
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Ingested.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentDescriptor' }
        '413': { $ref: '#/components/responses/LimitBreach' }
        '415': { $ref: '#/components/responses/UnsupportedInput' }
        '422': { $ref: '#/components/responses/UnprocessableInput' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /documents/{documentId}:
    delete:
      tags: [documents]
      summary: Purge a document session
      description: |
        Erases the document: upload bytes, canonical text and chat context, and with
        them every run of that document, its findings and its attempts. A finding
        carries the offsets, boxes and legal locators saying where in the contract the
        system looked, so an erasure control that kept them would mean "forget the
        text" rather than what it says. Every task reading the document is cancelled
        first -- analyses, and the routing and explaining calls a chat message starts,
        which hold the session and carry its text in their grounding -- so nothing goes
        on working over material the reader has been told is gone, and nothing files
        findings against records being removed. The browser sends this best-effort on
        close.

        Idempotent: a document the process no longer holds is already purged, so the
        response is 204 whether or not it was there. Retention sweeps and restarts drop
        the text while a browser tab still names it, and an erasure control must not
        report failure in exactly that case.

        Refused only while the final evaluation batch is sealed, when the records this
        would erase are the measurement in progress.
      operationId: deleteDocument
      parameters: [{ $ref: '#/components/parameters/DocumentId' }]
      responses:
        '204': { description: Erased, or already absent. }
        '409':
          description: |
            The final evaluation batch is sealed, so the runs this would erase are
            the measurement in progress (evaluation_batch_open).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/InvalidRequest' }
  /documents/{documentId}/content:
    get:
      tags: [documents]
      summary: Ephemeral canonical blocks and anchors
      description: |
        Returns canonical text blocks and their anchors for rendering. TXT and Word inputs
        render these blocks. For PDF the original file stays in the browser as a local File
        object; PDF.js renders it and uses the returned boxes for overlays. The server does
        not claim converted Word text preserves page layout.
      operationId: getDocumentContent
      parameters: [{ $ref: '#/components/parameters/DocumentId' }]
      responses:
        '200':
          description: Canonical content.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentContent' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/InvalidRequest' }
        '410': { $ref: '#/components/responses/ContentExpired' }
  /runs:
    post:
      tags: [runs]
      summary: Start one arm
      description: |
        Every arm accepts every document. ON supplies resolved internal-reference
        context where the document has any; where it has none the run is
        byte-identical to MID, which is a real experimental condition rather than an
        error. context_edge_count on the run record says which of the two happened.
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [document_id, arm]
              properties:
                document_id: { type: string, format: uuid }
                arm: { $ref: '#/components/schemas/ArmCode' }
      responses:
        '202':
          description: Accepted; poll the run or subscribe to its events.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/InvalidRequest' }
        '409':
          description: |
            Runs are concurrent by default. While the final evaluation batch is
            open the installation admits one active run (run_already_active).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '410': { $ref: '#/components/responses/ContentExpired' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /runs/{runId}:
    get:
      tags: [runs]
      summary: Run status, findings, telemetry and content availability
      operationId: getRun
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200':
          description: Run state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/InvalidRequest' }
  /runs/{runId}/cancel:
    post:
      tags: [runs]
      summary: Request a stop
      description: |
        Cancellation is a request. Attempt costs already incurred are retained, and units
        without a terminal finding become not_processed.
      operationId: cancelRun
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '202':
          description: Cancellation requested.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/InvalidRequest' }
        '409':
          description: Run already terminal.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /runs/{runId}/events:
    get:
      tags: [runs]
      summary: Server-sent event stream of status, counters and finding IDs
      description: |
        Events carry status, counters and finding IDs only. They never carry document or
        corpus text. Durable replay is not promised: on reconnect, fall back to GET
        /runs/{runId} for authoritative state.

        The server adds run_id at serialisation time; it is not a field of RunEvent
        in the codebase.
      operationId: streamRunEvents
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200':
          description: |
            An event stream. Each event is one RunEvent as JSON in the data field.
            The server injects run_id into each serialised event.
          content:
            text/event-stream:
              schema: { $ref: '#/components/schemas/RunEvent' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/InvalidRequest' }
...
components:
  parameters:
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: { type: string, format: uuid }
    RunId:
      name: runId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ContentExpired:
      description: |
        Content was purged by deletion, session expiry or process exit. Metadata and arm
        metrics for past runs remain readable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UnsupportedInput:
      description: |
        Unsupported file extension. Format is chosen by the filename extension of the
        multipart part (.txt, .docx, .doc, .pdf), not by media type.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UnprocessableInput:
      description: Empty, corrupt, or not canonicalisable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InvalidRequest:
      description: |
        The request body or a path parameter failed validation. Every 422 on this API
        carries the Error schema: a RequestValidationError handler replaces FastAPI's
        default body so one status never carries two shapes, and the submitted value is
        never echoed back.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    LimitBreach:
      description: Input exceeds a configured, frozen bound.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    DependencyUnavailable:
      description: A required local dependency is unavailable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message_pl]
      description: |
        Stable machine code plus Polish user text. The code is for the client; the text is
        for the person. Never contains document content.
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message_pl: { type: string }
        detail:
          type: object
          additionalProperties: true
          description: Bounded structured context. Never document or corpus text.
...
    ArmCode:
      type: string
      enum: ["off", "mid", "on"]
      description: |
        Mechanism names, not quality levels. The client labels are fixed:
        OFF — cały dokument
        MID — jednostki
        ON — jednostki + odwołania
    ReadMode:
      type: string
      enum: [native_pdf, ocr_pdf]
...
    CounterName:
      type: string
      description: |
        The counters a run emits. A free string here is what let the browser test double
        drift to "units_processed" while the server emitted "processed_units" with nothing
        to catch it; naming them makes a mismatch a generated-type error.
      enum:
        - processed_units
    ReferenceType:
      type: string
      enum:
        - full_internal
        - relative_internal
        - directional
        - annex
        - whole_document
        - external_act
    ReferenceStatus:
      type: string
      enum:
        - resolved
        - target_does_not_exist
        - within_unit
        - ambiguous_directional
        - outside_input
        - whole_document
        - external_act
    ReferenceRecord:
      type: object
      required: [citing_unit_id, reference_type, status, raw_text]
      description: |
        One cross-reference detected in a unit. Every unresolved status is retained.
        Only resolved depth-one internal targets contribute graph context.
      properties:
        citing_unit_id: { type: string }
        reference_type: { $ref: '#/components/schemas/ReferenceType' }
        status: { $ref: '#/components/schemas/ReferenceStatus' }
        raw_text: { type: string }
        target_unit_id: { type: [string, 'null'] }
...
    SourceAnchor:
      type: object
      required: [start_offset, end_offset]
      properties:
        start_offset: { type: integer, minimum: 0 }
        end_offset: { type: integer, minimum: 0 }
        read_mode:
          oneOf:
            - { $ref: '#/components/schemas/ReadMode' }
            - { type: 'null' }
        block_id: { type: [string, 'null'] }
        line_start: { type: [integer, 'null'] }
        line_end: { type: [integer, 'null'] }
        page: { type: [integer, 'null'] }
        bbox:
          type: [array, 'null']
          minItems: 4
          maxItems: 4
          items: { type: number }
          description: |
            PDF page points, top-left origin. Native and OCR anchors share this space;
            OCR boxes are divided by the render scale before being reported.
    ConversionProvenance:
      type: object
      required: [converter, converter_version, output_hash]
      properties:
        converter: { type: string }
        converter_version: { type: string }
        output_hash: { type: string }
    DocumentDescriptor:
      type: object
      required: [id, content_hash, unit_count]
      properties:
        id: { type: string, format: uuid }
        content_hash: { type: string }
        read_mode:
          oneOf:
            - { $ref: '#/components/schemas/ReadMode' }
            - { type: 'null' }
        conversion:
          oneOf:
            - { $ref: '#/components/schemas/ConversionProvenance' }
            - { type: 'null' }
        unit_count: { type: integer, minimum: 0 }
        expires_at:
          type: [string, 'null']
          format: date-time
          description: Bounded session expiry. After this, content returns 410.
    DocumentContent:
      type: object
      required: [document_id, blocks, units, references]
      properties:
        document_id: { type: string, format: uuid }
        read_mode:
          oneOf:
            - { $ref: '#/components/schemas/ReadMode' }
            - { type: 'null' }
        blocks:
          type: array
          items:
            type: object
            required: [id, text, anchor]
            properties:
              id:
                type: string
                description: >
                  Stable within one document. Readers that name their blocks keep
                  the name; the rest are addressed by position. A finding's
                  block_id refers to this.
              text: { type: string }
              anchor: { $ref: '#/components/schemas/SourceAnchor' }
        units:
          type: array
          items: { $ref: '#/components/schemas/Unit' }
        references:
          type: array
          description: |
            Cross-references detected across all units. The ON arm uses resolved
            references for graph context; all statuses are retained.
          items: { $ref: '#/components/schemas/ReferenceRecord' }
    Unit:
      type: object
      required: [id, text, anchor]
      properties:
        id: { type: string }
        text: { type: string }
        anchor: { $ref: '#/components/schemas/SourceAnchor' }
...
    SynthesisGroup:
      type: object
      required: [title, summary, finding_ids]
      properties:
        title: { type: string }
        summary: { type: string }
        finding_ids:
          type: array
          items: { type: string }
          description: >-
            Identifiers of findings on this same run. A grouping citing anything else is
            refused as an unusable model response, so each one resolves against findings.
    RunStatus:
      type: string
      enum: [running, completed, failed, cancelled]
    EventKind:
      type: string
      enum: [status, counter, finding, worksheet]
...
    Run:
      type: object
      required:
        - id
        - document_id
        - arm
        - status
        - measurement_valid
        - content_available
        - findings
        - metrics
      properties:
        id: { type: string, format: uuid }
        document_id: { type: string, format: uuid }
        arm: { $ref: '#/components/schemas/ArmCode' }
        status: { $ref: '#/components/schemas/RunStatus' }
        measurement_valid:
          type: boolean
          description: False for demo-mode runs and for any run whose parity is broken.
        parent_run_id: { type: [string, 'null'], format: uuid }
        interaction:
          oneOf:
            - { type: string, enum: [ask, contest, analyse] }
            - { type: 'null' }
        interrupted:
          type: boolean
          description: |
            True when a budget boundary was crossed. The client shows a critical
            document-level banner with units_not_processed.
        interruption_reason:
          oneOf:
            - type: string
              enum: [wall_time]
            - type: 'null'
          description: Persisted boundary that caused an interrupted run.
        content_available:
          type: boolean
          description: False once the document session is purged; content and chat return 410.
        requested_model: { type: string }
        returned_model: { type: [string, 'null'] }
        input_hash: { type: string }
        prompt_bundle_version: { type: string }
        corpus_snapshot_id: { type: string }
        config_version: { type: string }
        tool_bundle_version: { type: string }
        parameters:
          type: object
          additionalProperties: true
        retry_policy: { type: string }
        concurrency: { type: integer }
        wall_budget_seconds: { type: number }
        created_at: { type: string, format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
        error:
          oneOf:
            - { $ref: '#/components/schemas/Error' }
            - { type: 'null' }
        findings:
          type: array
          items: { $ref: '#/components/schemas/Finding' }
        synthesis:
          oneOf:
            - type: array
              items: { $ref: '#/components/schemas/SynthesisGroup' }
            - type: 'null'
          description: >-
            Themed grouping of this run's findings, produced by the synthesis role and
            held for the same retention window as the document text. Null before the
            role has run, once the window closes, and on an installation that keeps no
            document text at all.
        metrics:
          oneOf:
            - { $ref: '#/components/schemas/RunMetrics' }
            - { type: 'null' }
    RunEvent:
      type: object
      required: [kind]
      description: |
        Carries status, counters and finding IDs only. Never text. The server adds run_id
        at serialisation; it is not a field of RunEvent in the codebase.
      properties:
        kind: { $ref: '#/components/schemas/EventKind' }
        status:
          type: [string, 'null']
          description: Free snake_case status string (kind=status events only).
        counter_name:
          oneOf:
            - { $ref: '#/components/schemas/CounterName' }
            - { type: 'null' }
          description: Which counter this event carries (kind=counter events only).
        count:
          type: [integer, 'null']
          description: Counter value (kind=counter events only).
        finding_id:
          type: [string, 'null']
          format: uuid
          description: Finding UUID (kind=finding events only).
        role:
          type: [string, 'null']
          description: >-
            Which role wrote the worksheet entry (kind=worksheet events only): one of
            researcher, analyst, verifier, user, system. Never what it wrote.
        entry_kind:
          type: [string, 'null']
          description: Kind of worksheet entry appended (kind=worksheet events only).
        unit_id:
          type: [string, 'null']
          description: Call unit the worksheet belongs to (kind=worksheet events only).
...